feat: Basic functionality with several graphs
Created the basic layer of functionality. As of right now it's a cli app
that pulls information from the telegram api, saves it in the mongodb
and than performes as series of queries in order to acquire specific
information and create graphs. Graphs available right now are: activity
in different day times, number of messages per day over time, words
usages bar chart which shows how many times a word was used.